I changed code in SnakeGame and in SnakeExample.
I added in SnakeGame function that gives me coordinates of food AND creation of food, I edit randomizer to get food onli inside of board
And in SnakeExample I added part of code about direction. Now Snake always checks similarity in X and Y axis and if they are same tries to turn in right way to eat food.